% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{\%_\%}
\alias{\%_\%}
\alias{\%-\%}
\alias{\%.\%}
\alias{\%//\%}
\alias{\%&\%}
\alias{\%&&\%}
\alias{\%|\%}
\title{Concatenate with an underscore}
\usage{
a \%_\% b

a \%-\% b

a \%.\% b

a \%//\% b

a \%&\% b

a \%&&\% b

a \%|\% b
}
\description{
\%_\% pastes two strings by "_".

\%-\% pastes two strings by "-".

\%.\% pastes two strings by ".".

\%//\% pastes two strings by "/".

\%&\% pastes two strings by "".

\%&&\% pastes two strings by " ".

\%|\% pastes two strings with a "|".
}
