% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SWATdata}
\alias{SWATdata}
\title{SWAT model setups for \code{SWATplusR}}
\arguments{
\item{a, b}{characters.}
}
\description{
\code{SWATdata} provides SWAT model setups for SWAT2012 and different SWAT+
revisions for the gauge J of the Little River Experimental watershed (LREW,
Bosch et al. 2007).
}
\details{
The following datasets are currently available from \code{SWATdata}:
  \tabular{rccc}{
  \strong{Version} \tab \strong{Windows} \tab \strong{Linux} \tab \strong{Spatial} \cr
   SWAT2012 Rev622 \tab X \tab   \tab X\cr
   SWAT2012 Rev637 \tab   \tab X \tab X\cr
   SWAT+    Rev 57 \tab X \tab   \tab X\cr
     }
}
\references{
\cite{ Bosch, D. D., Sheridan, J. M., Lowrance, R. R., Hubbard, R. K.,
       Strickland, T. C., Feyereisen, G. W. and Sullivan, D. G.: Little river
       experimental watershed database, Water Resources Research, 43(9),
       \url{https://doi.org/10.1029/2006wr005844}, 2007}
}
